% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil20.R
\docType{package}
\name{coil20-package}
\alias{coil20}
\alias{coil20-package}
\title{coil20: Download COIL object image datasets}
\description{
Helper functions to download the processed COIL-20 and COIL-100 object
image datasets and convert the PNG files into row-based R objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jlmelville/coil20}
  \item Report bugs at \url{https://github.com/jlmelville/coil20/issues}
}

}
\author{
\strong{Maintainer}: James Melville \email{jlmelville@gmail.com}

Authors:
\itemize{
  \item James Melville \email{jlmelville@gmail.com}
}

}
