% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil20.R
\name{download_coil100}
\alias{download_coil100}
\title{Download COIL-100}
\format{A data frame with 49153 variables (128 x 128 pixels for each
of the three channels):

\describe{
\item{\code{r_x1_y1}, \code{r_x1_y2}, ... \code{r_x1_128}, \code{r_x2_y1} ...
  \code{r_x128_y128}}{Pixel values in the red channel ranging from 0 to 1.
  The \code{x} index increases from left to right, and the \code{y} index
  increases from top to bottom, i.e. \code{r_x1_y1} is top left and
  \code{r_x128_y128} is the bottom right.}
\item{\code{g_x1_y1}, \code{g_x1_y2}, ... \code{g_x1_128}, \code{g_x2_y1} ...
  \code{g_x128_y128}}{Pixel values in the green channel ranging from 0 to 1.
  The \code{x} index increases from left to right, and the \code{y} index
  increases from top to bottom, i.e. \code{g_x1_y1} is top left and
  \code{g_x128_y128} is the bottom right.}
\item{\code{b_x1_y1}, \code{b_x1_y2}, ... \code{b_x1_128}, \code{b_x2_y1} ...
  \code{b_x128_y128}}{Pixel values in the blue channel ranging from 0 to 1.
  The \code{x} index increases from left to right, and the \code{y} index
  increases from top to bottom, i.e. \code{b_x1_y1} is top left and
  \code{b_x128_y128} is the bottom right.}
\item{\code{Label}}{The id of the object represented by the image, in the
  range 1-100. Stored as a factor.}
}

The name of each row in the data frame is in the form "<obj>_<angle>", where
<obj> is the id of the object (numbered from 1 to 20) and <angle> is the
viewing angle in degrees (numbered from 0 to 355). e.g. "5_150" is the fifth
object viewed at an angle of 150 degrees.}
\usage{
download_coil100(file = tempfile(), cleanup = TRUE, verbose = FALSE)
}
\arguments{
\item{file}{name of file to download to. Defaults to \code{\link{tempfile}}.}

\item{cleanup}{If \code{TRUE}, then delete any downloaded files when
finished.}

\item{verbose}{If \code{TRUE}, then download progress will be logged as a
message.}
}
\value{
Data frame containing the COIL-100 database.
}
\description{
Download processed COIL-100 database of object images.
}
\details{
Downloads and expands the zip file of the processed COIL-100 database of
object images and converts them to a data frame. There are 100 objects, each
represented by 72 poses, in an RGB PNG format. Each image has a
resolution of 128 x 128.
}
\note{
This function requires the \code{png} package
\url{https://cran.r-project.org/web/packages/png/} to be installed.
}
\seealso{
For more information see
 \url{http://www.cs.columbia.edu/CAVE/software/softlib/coil-100.php}.
}
