% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil20.R
\name{show_object}
\alias{show_object}
\title{Visualize COIL-20 image.}
\usage{
show_object(df, object, pose)
}
\arguments{
\item{df}{Data frame containing COIL-20 objects.}

\item{object}{Index of the object to display. Should be an integer between
1 and 20.}

\item{pose}{Index of the pose to display. Should be an integer between
0 and 71.}
}
\description{
Display a COIL-20 object pose.
}
\examples{
\dontrun{
# show the fourth pose of the fifth digit
show_digit(coil20, object = 5, pose = 4)
}
}
